+++ protocols/kad/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Update to `libp2p-swarm` `v0.34.0`.
 
-- Require owned key in `get_record()` method (see [PR 2477])
+- Require owned key in `get_record()` method (see [PR 2477]).
 
 [PR 2477]: https://github.com/libp2p/rust-libp2p/pull/2477