+++ protocols/kad/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.35.0 [unreleased]
+
+- Update to `libp2p-core` `v0.32.0`.
+
+- Update to `libp2p-swarm` `v0.34.0`.
+
 # 0.34.0 [2022-01-27]
 
 - Update dependencies.